FERC plantid training notebook and Classifier
Added some helper functions to the FERCPlantClassifier class to help
support the model training/fitting, adapted from the sklearn notebook.

Also created a new notebook with a couple of examples of how to manually
categorize FERC plants, so we can create a training set to work with.

Partially address issue #144