Reorganize output and analysis modules into subpackages
The output.py module had grown to ~1500 lines and had become unwieldy.
It also wasn't entirely clear how it was organized and meant to be used.
As we've gotten more collaborators and users looking at the code, I
wanted to make it clearer that there's an output layer for grabbing more
dataframe/spreadsheet like compilations of data, rather than having
everyone try to interface directly with the database (though of course
it's still there.)

Separating the output functions by data source also makes it cleaner to
add new output functions for new data sources -- as we now need to do
for the EPA CEMS data.

This commit also creates a module which does nothing but define the
PudlTabl class, which is used to compile TABuLar outputs for use or
export -- these are either more spreadsheet-like versions of individual
data tables, or compilations of derived values, as in the case of the
MCOE calculations.

Export functions (of which there's currently only one, for spreadsheets)
are in a separate output.export.py module.

Analysis routines like the MCOE have been separated into their own
subpackage at the top level, and are referenced by the PudlTabl class.
Several widely used functions were extracted from the analysis.py module
and put into a top level helpers.py module (I would have called it
utils.py but we have so many other things called utils already I thought
that would be confusing...), and left the rest of the (mostly obsolete)
analysis functions in analysis.analysis... which still needs to be
culled and re-named and/or split up.

The affected test modules and transform modules are also included.  This
is an entirely refactoring commit -- really no new code here.  Though,
many new docstrings.