How to add a datasource!
Whoo. Closes Issue #177. This is a draft in many ways and we aresuper
willing to hear feedback from folks and modify standards, but thisshould
give a general outline of the pieces of the etl process and howwe've
thought about them and want them implemented moving forward. Thereare a
few sections (such as glue) that are partially aspirational. Butwe'll
move towards implementaion.