Compartmentalizing ETL by data source.
Using the list of years given as an argument for init_db, now each
datasources' ETL process can now be easily broken out between
datasources and the glue between ferc and eia is conditionally ingested.
A new table is created called datasets, which should be used in output
functions and any analysis to determine wether a dataset is in the
database. Still a long way to go on this... i.e. conditional importing
of the models modules so PUDL won't have a bunch on empty tables when
ingested. Working on Issue #195.