Re-organize FERCPLantClassifier and helper functions.
Got basic versions of fit() and predict() working -- storing the entire
cosine similarity matrix inside the object, and requiring the original
dataframe of information about the plants to be passed in as part of the
__init__() function, which is obviously not idea -- but for this problem
it seems like we have almost nothing *but* hyperparameters to vary. And
the distance metric that results in the categories kind of depends on
all of the records being available. Which is fine, but not the way these
things normally work.

Have still left out the single numerical variable, since it requires
reshaping to work in the Pipeline. It's all so ugly right now. Blargh!

Working toward issue #144