Creating a configuration file.
This still need a little jiggering to get final settings and placement 
but I created a configuration file (settings.yml) which gives users the 
ability to easy ability to denoate which years and tables that they want 
to import into pudl. The settings.yml file is loaded in config.py, which 
also includes other settings like directory paths that are settings but 
we don't really want users to change.

The tests still reference the working years/tables, with the exception 
of CEMS, which references the constants.

Things left to do: Move the the csvdir setting into the yml file. Add 
ability to specify configuration file through command line argument... 
and there is probably more.