Update script to use new SETTINGS dictionary
With the re-organization of many of the PUDL package-wide settings into
the SETTINGS dictionary, I needed to change where some of the directory
values were coming from.

Also moved the packaging scripts up on level in both the scripts and the
results hierarchy -- since every data package will have to have a unique
name, which will include the data source, it seemed silly to insert an
additional level in the directories (and honestly, maybe we should do
this for the datastore too).