Simplify FERC Plant Classifier code
In scikit-learn 0.20 (which is currently at the release candidate phase,
and expected to be released in the fall of 2018) functionality is added
to integrate much more directly with data that has been preprocessed
and stored in pandas data frames. See for example:
 * [this tutorial on the new features](https://medium.com/dunder-data/from-pandas-to-scikit-learn-a-new-exciting-workflow-e88e2271ef62)
 * [Column Transformer with Heterogeneous Data Sources](http://scikit-learn.org/dev/auto_examples/compose/plot_column_transformer.html)
 * [Column Transformer with Mixed Types](http://scikit-learn.org/dev/auto_examples/compose/plot_column_transformer_mixed_types.html)
 * [Column Transformer documentation](http://scikit-learn.org/dev/modules/generated/sklearn.compose.ColumnTransformer.html#sklearn.compose.ColumnTransformer)

As a result, a bunch of the hacked mess that was taking up space
supporting the FERC Plant Classifier has been removed.

The final `transform` function was also moved up into the main section
of the module, where the other standardized transformation stuff lives,
to keep it separated from the scikit-learn stuff happening lower down.