Removed pinned versions which means Python 3.7
Some of our pinned versions were pretty out of date. I removed all the
pinned versions and tested, and everything still works, with the newest
stable versions of all the packages we depend on.

Removed the pip install of scikit-learn 0.20rc1 because the official
0.20 release happened.

The biggest change in removing *all* of these pinned versions is the
fact that the environment now uses python 3.7, which is the current
stable python release.