Removed some invalid records from training set.
Unfortunately some of the records which had been included in the human
identified time series were records being dropped from the DB in the
initial data cleaning. In particular, some of them were associated with
plants that reported zero net generation.  This was blowing up some
other parts of the classification system, and so those individual record
IDs were removed from the training set.  This is not the ideal
solution...