Renamed data package dirs to conventional names
This isn't perfect, but it's a little closer to what we've been doing
elsewhere.

pudl-ferc1 needs to be distinct from pudl-ferc714 or pudl-ferceqr, and
matches the naming we're using elsewhere.

pudl-epacems matches the naming we're using elsewhere.

pudl-eia is so named because it includes both 923 and 860 data, which
are pretty much inseparable in our system. Not sure what we would do if
we packaged up other EIA data -- maybe need to think about this a bit
more. There doesn't appear to be a way to relate tabular data packages
to each other (though you can relate tabular data resources to each
other within a given data package).