Add explicit module imports to top level package
For some reason, we've never explicitly added the list of modules which
we want to have access to when pudl is imported at the top level. How
and why anything was working before this escapes me -- name space and
importing is still kind of a mystery to me -- but with the addition of
these module imports, all you need to do is `import pudl` (assuming the
top level package is in your system path) and you get all of the listed
modules imported as well.  Note that it is *not* everything -- there are
still plenty of things that aren't ready for general use laying around.