Make all FERC plant names canonically lower case
Previously we've been converting all FERC plant names to Title Case to
improve inter-year consistency of the names (as well as stripping
leading and trailing whitespace, and compressing internal whitespaces).
However, Title Case is a little bit non-deterministic, because it's
affected by whitespace which may or may not be present, given the
variety of plant names reported in FERC. This ends up mattering in the
FERC plant ID assignment, since strings with different cases are
different strings. Up until now we've been converting them all to lower
case for the FERC plant ID assignment, but this means we end up having
plant strings which are Title Case in some places and lower case in
others, and that's confusing. This commit forces all the FERC plant
names to be lower case consistently.

Closes #219