Integration of plant_id_ferc1 assignment into FERC ETL
Updated the plant_id_ferc1 assignment process to work on the
plants_steam_ferc1 proto-table (the dataframe as it is at the end of the
current transform step), since that's where the ID assignment will
ultimately need to take place.

This included swapping out the CPI plant type assignments, and reverting
to using our own (which have more categories, and are thus likely more
useful for matching plant records together).  The CPI plant types have
been preserved in the ever growing constants module.

Changed the names of the columns that are used in the FERC Plant
Classifier class.  Update pre-processing of some of those columns in the
transform step to ensure they work with sklearn module.

Removed the ad-hoc FERC plant cleaning function which was hanging out
with the FERC Plant Classifier in the transform module, since it now
uses the canonical proto-table as input.

Progress toward #219 and #144