Assign FERC Plant IDs based on record similarity
At long last, the process for identifying related records in the FERC
plant table has been integrated into the ETL process.

There's still tweaking to be done with the values of the parameters
which are used to determine the related groups, and the remaining ~10%
of all records which aren't assigned to any FERC plant (but which are
still valid records) need to be dealt with. Refinements are listed as
other issues. This commit makes progress on #144 and closes #219