Integrate plant_id_ferc1 into fuel_ferc1 table
During the FERC Form 1 transform process, the plant_id which is
generated and assigned in the steam_plants_ferc1 table is now inserted
into the fuel_ferc1 table, by merging on the shared fields that are
co-reported to FERC in the two tables: report_year, utility_id_ferc, and
plant_name -- in this and only this instance, it appears that the plant
name can serve as part of a foreign key -- the fuel and steam tables get
the exact same values.

Note that this merge *need not* result in the same plant_id_ferc1 being
assosicated with records having the same plant name and utility ID.
Those fields are just used to connect the FERC steam and fuel tables,
bringing across whatever ID has been assigned on a record-by-record
basis in the plants_steam_ferc1 ETL process.

This commit closes #219 and makes progress toward #144.