Update ETL to include 2017 EIA 860 & 923
The EIA Form 923 and 860 data has been released for 2017. It is now
available to the ETL process. 2017 has been added to the list of working
years for eia860 and eia923 in constants.py, and in the commented out
lists of all working years in the default settings.yml file that the
init_pudl.py script uses to initialize the database.

In addition the change to the 2017 download URLs from EIA for 860 and
923 have been addressed in the datastore module, so update_datastore.py
now works for those sources.

The EPA CEMS data for 2017 is also now included. No additional work was
required beyond adding 2017 to the list of working years.

The FERC Form 1 data for 2017 still needs to be integrated (Issue #224),
but that's turning out to be a little more challenging.

Closes #225 (Integrate 2017 EIA 860 data)
Closes #226 (Integrate 2018 EIA 923 data)
Closes #231 (Update datastore.py to deal with new EIA URLs)