Fix the weird pre-2013 year requirement for EIA 860
Previously, it was necessary to have at least one year from before 2013
included in the years of data loaded for the EIA860, because otherwise
not all of the columns defined in the database, and manipulated in the
transform step were present -- the columns of EIA860 have changed a lot
over the years.

This commit removes this requirement, by concatenating any missing
columns onto the EIA860 dataframes before they are returned from the
extract functions, ensuring that the extracted dataframes always have
all the columns.

Additionally, the types of some columns were not being correctly set to
integer, and those were fixed -- we still don't have a
from-the-beginning assignment of types anywhere that I can see, which
seems bad.  Also, lots of room for consolidation of repeated cleanup
code in there with a few short helper functions.

Sadly, after fixing this bug... I discovered another one. If only FERC
2016 is loaded, the fuel type cleanup in fuel_ferc1 blows up in an
infinite recursion. So weird.

Closes #197