Make cleanstrings work with no leftovers
Calls to cleanstrings where there were no leftover bad strings unmapped
were resulting in an infinite recursion (since it was being asked to
replace nothing with nothing...). This was causing the transform step of
the FERC Form 1 to fail when only 2016 was included, since the stringmap
completely maps all of the strings present in 2016. Cleanstrings now
checks to see whether there are any leftover bad strings that need to
be assigned the "unmapped" value, and only attempts to do that
replacement if there are.

Closes #232