Removed unused yearly_to_monthly_eia923 function
There were two slightly different versions of the
yearly_to_monthly_eia923() function, one in `pudl.extract` and one in
`pudl.transform`. Only the one in `pudl.transform` was being used. I
deleted the unused version, and renamed it to
_yearly_to_monthly_records() to remove the name redundancy and make it
clear that this function is only used within this module. Also changed
the coalmine_cleanup() function to _coalmine_cleanup() and switched a
few strip() and lower() transformations to use
`pudl.helpers.strip_lower()`

Closes #233