Comment out script deployment. Set minimum SQLAlchemy version.
We have several scripts which we'll want to deploy at installation
eventually... but they aren't ready for that just yet. To avoid
confusing folks and having them show up in your path, I've commented
them out for the time being.

Also set a minimum version for sqlalchemy of 1.2, since we are now using
the Column(comment="") parameter to annotate database fields, and it's
not available before 1.2