Annotate FERC columns, remove simple glue tables
In order to minimize the number of places we need to go looking for
metadata when creating data packages, I've started making use of the
Column.comment attribute in the database table definitions. This commit
adds comments to the fuel_ferc1 and plants_steam_ferc1 tables.

In the process of making those changes... it became obvious that there
were a bunch of small glue tables that were either entirely unused, or
which would be more appropriate as simple Enum() types rather than whole
tables. So I made those changes.

Also further simplified and rationalized some of the codes which were
being used. E.g. in the coalmine_eia923.state column, rather than using
a jumble of non-standard 2-letter country codes for international mines
(because the official 2-letter country codes clash with some of the
2-letter US state abbreviations, e.g is CA California or Canada?), I
swapped in the 3-letter ISO-3166-1 country codes for other countries.