Replace default parameter lists with tuples
It turns out that it's dangerous to use lists as default parameters in
function calls -- especially lists that come from somewhere else --
because what you're passing is a reference to that faraway list.  If you
modify the list in the function, the "original" version of it also gets
modified... which is bad.

I've replaced a bunch of the lists we're using this way in the constants
module with tuples instead, which are immutable, hopefully avoiding this
failure mode.