Update fix_int_na() & add DB table dependency finder
Previously fix_int_na() has taken a Series (a DataFrame column) and
returned a Series, with the columns fixed. This means when there's more
than one column to be fixed, it has to be called multiple times.  It
also means you can't use .pipe() to chain fix_int_na() with other
methods. Now it takes a dataframe and a list of columns to be fixed,
which is a bit cleaner. astype() and replace() also operate on a whole
dataframe with appropriate arguments, so the function is still very
small. I updated all the uses of fix_int_na I could find to use the new
syntax.

I also added a couple of functions to pudl.helpers that figure out which
other tables a given input table (or list of tables) depends on, for use
in creating data packages which are self-contained database
specifications + the data that goes inside.

Progress toward #211