Add DB comments to FERC tables
Added comments to all the FERC tables and columns, for use as metadata
in the data packages (and also general self-documentation.  In the
process, found a variety of inconsistencies between different FERC
tables (naming of columns, incorrect types, etc.) and fixed them. This
included effective cleanup of the purchased power table, which was still
pretty raw. Lots more mess to be found in the FERC tables, as always.

Progress toward Issue #211