Deduplicate comments in different FERC tables
Many FERC tables have identical fields (utility_id_ferc1, report_year,
etc.) and it seemed redundant and harder to maintain to have them all be
commented with the same independent string, so I made a collection of
comment string variables near the top of the module that are re-used as
appropriate throughout the Column definitions. There are more that could
be added, but I think I got all the most frequently re-occurring fields.