Havrest values for plant entity tbl
Working on Issue #135 and Issue #250. For #135, all 'static' plant 
records are now being harvested into the plant entity table and have 
been extracted from their native tables. To decide which record to grab 
from many tables, I've created a process to determine how consistent a 
value is reported. If a value matches 70% of the time, the value is 
choosen to end up in the entity table. Right now, the latitude and 
longitude are the only columns that need a special process applied to 
them.

For #250, the lat/long are now all assiciated with plant records in 
plant_entity_eia. Because the annual records were not completely 
consistent, I rounded the inconsistent records a bit in order to match 
more - this is sort of a hack. We should eventually change this to find 
the closet actual location although this shouldn't effect timezones.