Add routine to aggregate fuel by FERC plant
Each record in the FERC Form 1 corresponds to a particular type of fuel.
Many plants -- especially coal plants -- use more than one fuel, with
gas and/or diesel serving as startup fuels. In order to be able to
classify the type of plant based on relative proportions of fuel
consumed or fuel costs it is useful to aggregate these per-fuel records
into a single record for each plant.

Fuel cost (in nominal dollars) and fuel heat content (in mmBTU) are
calculated for each fuel based on the cost and heat content per unit,
and the number of units consumed, and then summed by fuel type (there
can be more than one record for a given type of fuel in each plant
because we are simplifying the fuel categories). The per-fuel records
are then pivoted to create one column per fuel type. The total is summed
and stored separately, and the individual fuel costs & heat contents are
divided by that total, to yield fuel proportions.  Based on those
proportions and a minimum threshold that's passed in, a "primary" fuel
type is then assigned to the plant-year record and given a string label.

Also required for FERC non-fuel OpEx regressions in #204