Accommodate file name changes from 2018 EIA 923 data
A version of the full year of 2018 EIA 923 data has been released, and
so they have updated the file names, moving the 2017 data into the
"archive" folder (ugh, why do they do this). This commit works around
that, and hopefully makes the datastore management a little bit more
robust to this kind of change in the future.