Add an example notebook validation test.
For a while we have been wanting to integrate visual data validation
tests that can be automatically run. This commit adds a notebook
test directory (test/notebooks) and a first notebook, that compares
several variables between FERC and EIA plants as an example. It still
needs to be integrated into a test script somehow...

This commit also updates some of the environment requirements, including
the nbval pytest plugin.