Integrate fuel fractions into FERC Plant ID assignment
This commit integrates information about the relative proportions of
various fuels consumed into the feature matrix used to associate
plant records from different years with each other.

It also removes the plant_id_ferc1 field from the fuel_ferc1 table,
since it cannot be uniquely and completely merged in, given the lack of
unique mapping between plant and fuel table records (due to year,
utility, plant_name duplications... which are differentiable on the
basis of other information in the plant table).

Also simplifies and rationalizes the ColumnTransformer that's part of
the data preprocessing Pipeline for the FERCPlantClassifier.

This marginally improves the recovery of record associations, and could
probably be further tweaked, but it's in there structurally now at
least.

Closes: #266