Add assertion into EIA harvesting
The assertions are here to ensure that the harvesting process is 
producing enough consistent records. When every year is being imported 
the lowest consistency ratio should be .97, with the exception of the 
latitude and longitude, which has a ratio of ~.94. The ratios are better 
with less years imported. Closes Issue #135.