Add FERC fuel use by plant output routines
Addresses issue #265 by adding an output routine for FERC Form 1 that
compiles fuel use on a per-plant basis (as indicated by plant name and
utility ID) allowing fuel usage and primary fuel information to be
used with the FERC steam plant table / output. Unfortunately there was
no unique correspondence between the steam_plants_ferc1 and fuel_ferc1
tables, so this kind of summary is necessary. Makes use of the
fuel_by_plant_ferc1() function introduced in the FERC Form 1 transform
module, and applies that function to the post-transform FERC Fuel data
as stored in the DB.

Adds the necessary method to the PudlTabl output object as well.

Closes #265