Improve heat rate function name consistency
The heat rate functions in the PudlTabl output objects were not named
using the same convention as the other tabular output functions. Instead
of using the same short name (hr_by_unit and hr_by_gen) as the
dataframes used to store that data in the object, they used the longer
names found in the mcoe module. Now they match the local data members
for consistency.