Implement logging, expand FERC tables, test datastore
This is a gigantic commit covering many different issues, which is extremely
bad practice and I apologize in advance if it makes a mess!

### Implement Logging:
We have been accumulating ever more `print()` statements throughout the code,
and I finally couldn't take it any more. I have replaced them all with calls
to the standard Python logging module, which gives us much better control
over what is seen when, without needing to pass around "debug=True" or verbose
flags (which have also been removed en masse from all the functions, unless
they controlled something other than informational output).

This change to how the library does output will take a little getting used to,
but in the long run it will be worth it (I swear).

I categorized this work under the Maintenance project.

Closes #282

### Add datastore module tests:
Unfortunately in the course of switching over to the logging system, I somehow
broke the datastore update script (which I just happened to check on a lark,
since I was afraid I might have broken something that wasn't getting tested,
as the logging changes touch virtually every file.) It turns out that the
multi-threaded download function was failing (with a now obsolete verbose flag)
but not giving any kind of output indicating why.  In the process of fixing
this, I went ahead and created a test module for the datastore, which tries to
pull down both the oldest and the most recent years worth of data for each of:
ferc1, eia860, eia923, and epacems (albeit only for the state of Idaho, which
is the smallest of all the CEMS states, in terms of data). We really need to
expand the tests to cover more things... but this is a start at least!

Also in this vein, I added `scripts/pudl_test.sh` which outlines the pytest
commands one needs to run to do as good a job as possible of testing that
everything works, including any notebooks that are under `docs/notebooks` or
`test/notebooks`

Closes #205

### Expand universe of usable FERC Form 1 Tables / Years:
In a totally unrelated effort that definitely should not have overlapped with
the above work, I expanded the universe of FERC Form 1 years and tables that
can be cloned into the ferc1 database (NOT PUDL) by testing every combination
of year (1994-2017) and the 117 FERC Form 1 DBF tables. Many Just Work™. This
should not change the PUDL database at all for now, but it means that the FERC
database can populate with a lot more raw data, which we can bring over as need
be.

I replaced `pudl.constants.ferc1_working_years` with a dictionary of lists
called `pudl.constants.ferc1_working_table_years` enumerating all the years
that worked for each of the tables. Now you can try and load any of the tables
for any year that there's data, and if it's a known to be working table year,
it'll get cloned. If it's not known to be working, it will be skipped.

I updated `pudl.constants.ferc1_default_tables` to include the additional
tables of interest for Uday / RMI, and this necessitated updating the subset of
FERC DBF files that are included in the repository for use in our continuous
integration (Since either Travis CI or FERC has decided we should not be
downloading a full year of data for testing every time).

Someday we should debug why the various broken years aren't working, as it
doesn't seem to be big structural changes -- I suspect it's a bunch of small
poorly enforced database constraints that DBF lets you get away with, and
postgres does not. But, what do we say to the God of Death?  Not today!

Closes #68