Migrate FERC Form 1 DB to SQLite
In anticipation of moving to data packages (#258) and away from
postgresql as the canonical output from our data processing pipeline, I
swapped out the underlying database engine that receives the cloned FERC
DB when we pull it out of the DBF files. Now it uses SQLite. This will
allow us to get rid of postgresql as a requirement when the datapackage
switch happens.

Upsides:
 * SQLite is much faster than Postgres for cloning the DB. All working
   table-years from 1994 to 2017 can be cloned in less than 10 minutes
   on a fast laptop.
 * SQLite is also more robust to some of the foibles of the DBF files.
   more table-years successfully extract without any additional work.
 * Users will not have to deal with any kind of database access
   permissions during setup.
 * You don't have to worry about generating stale database connections
   that don't get cleaned up.
 * Because the whole DB is just a file on disk, you can turn the "test"
   database into the "live" database simply by re-naming the file, or
   set aside a fully cloned DB while you mess around with something
   instead of having to clobber it and start over next time.
 * The SQLite database file is self contained (though not particularly
   transparent) and can be shared with others directly.

Right now a maximally populated FERC Form 1 database takes up about
750MB on disk. This is... more than a factor of 10 smaller than the
original source data so... we must be missing something substantial
(maybe it's all the footnotes, that we discard at the beginning?).

Additional work will be required if anyone wants the missing tables and
years badly enough to debug the broken table-years (or pay us to do it).

The getting started documentation needs to be updated to correctly
reflect these changes or it'll confuse people badly.

A little function `pudl.analysis.analysis.check_ferc1_tables()` will
exhaustively try to clone one table-year at a time and report back on
its success or failure, if we need to regenerate the list in the future.
The results of that function are currently stored in
`pudl.constants.ferc1_working_table_years` which is referred to by the
`pudl.extract.ferc1` processes to avoid attempting to pull in bad
table-years.

Other than the initial Postgres database setup, this change should be
pretty transparent, since the database connections should all be getting
made using the `pudl.extract.ferc1.connect_db()` function, which now
uses a different database URL, pertaining to the SQLite DB. In keeping
with our general pattern, the FERC DB outputs can now be found at:
 * `pudl/results/sqlite/ferc1.db`
 * `pudl/results/sqlite/ferc1_test.db`

Closes #283