Adjust system to use a complete FERC Form 1 SQLite DB
Separate script to load complete FERC Form 1 DB:
 * `scripts/ferc1_to_sqlite.py`

Separate settings file for the FERC DB cloning script:
 * `scripts/settings_ferc1_to_sqlite_default.yml`

Renamed the settings file for init_pudl.py to be comparable:
 * `scripts/settings_init_pudl_default.yml`

Because the settings files are essentially required to run the scripts, I got
rid of the `-f` flag on the scripts -- they both simply require a settings file
as an argument.

Lots of updates required to get the FERC Extract step working (total re-write).
Also lots of changes to make the testing work appropriately.

The FERC Form 1 DB clones completely into a local SQLite database -- all tables
and all years (going back to 1994).  This is a useful thing in its own right,
as opposed to just being an input into PUDL, so it has been separated from the
PUDL ETL process a bit. You have to run the `ferc1_to_sqlite.py` script before
you can run the PUDL database initialization.

There are a couple of huuuuuge tables that are excluded from the cloning
process by default because they bloat the DB up by a factor of 10x in size,
and they mostly contain binary data that's unreadable:
 * `f1_footnote_tbl`
 * `f1_note_fin_stmnt`

But you can load them if you really want to.

Added `scripts/test_pudl.sh` more as a human-readable example of what one
needs to do to fully test the whole system (insofar as that's possible right
now...)

Still need to update the documentation to go along with all this. It's in
progress.

Closes #230
Closes #285
Closes #68 (again...)