Ensure transform retains all steam plant records
Previously, as seen in #242, we were losing a few steam plant records
every time we did the transform step... and not the same ones every
time, or the same number every time, which seemed weird and bad.

This was happening because we use a little bit of network analysis to
assign plant IDs to the various FERC plant records -- basically any pair
of records that makes up an "edge" within one of the interconnected
components of the graph of record associations (as with the
boiler-generator connections) gets labeled with the same
`plant_id_ferc1`.

In the Plant ID assignment, I was assuming that every record shows up as
being "connected to itself" (both a source and a target node in the
graph specification). This is *normally* true but... apparently not
*always* true. So in a dataframe with columns of (`source`, `target`,
`plant_id_ferc1`) I was dropping one of the source/target columns, and
then dropping duplicates, and assuming the remaining ID/node association
was exhaustive.  Which it wasn't quite.

Now the routine stacks the source only and target only records on top of
each other with their respective plant IDs, and drop duplicates, and
every single record is included.

Since we use a similar process for creating the EIA boiler generator
associations, I checked to make sure that the same bug doesn't exist
in `pudl.transform.eia` and thankfully it does not, because in the case
of the BGA, we have two types of nodes: boilers, and generators, which
can be connected to each other, but never to themselves, so we are
treating the nodes distinctly in the ID assignment process, and
purposefully assigning the `unit_id` to both the boilers and the
generators independently.

Closes #242