Move frictionlessdata python packages to pip install
Updated the `environment.yml` file that specifies the `pudl` conda
environment to install all of the frictionlessdata python packages using
`pip` instead of `conda` because there are some weird version
requirements in the dependencies on the `conda-forge` channel which end
up prohibiting the use of Python 3.7. I've been filing issues on the
appropriate frictionlessdata repos, but it seems like it might be a
little while until it's resolved, and there's no reason that PUDL
shouldn't work on both Python 3.6 and 3.7 when we finally make our first
release.