Removing the incorrect use of == True
I had two instances of selecting a portion of a dataframe based by using 
df[df['column'] == True] as opposed to df[['column']] for a column that 
was meant to be a boolean column. I had previously only assinged True 
(not False) so it wasn't actually a boolean column so I fixed that and 
had to modify some dependencies.