Make cleanstrings() usable in DataFrame.pipe()
Many of our data cleaning routines are *almost* perfect for pandas
method chaining, and could be much more compact an (in some cases at
least) readable if they were done that way. However, the
`pudl.helpers.cleanstrings()` routine was written before we knew how
method chaining worked, so it was incompatible.

This commit replaces `cleanstrings()` with a new function that takes a
dataframe and returns a dataframe, and can be used to clean up multiple
columns in one call.  It's a wrapper around `cleanstrings_series()`
which does the same thing, but for a pandas Series.

It also converts a couple of the `transform` module functions to use
method chains (basically... all the ones that were using `cleanstrings()`
and to make use of the new functions.

Closes #308