Beging the transtion from db init to exporting packages
Everything pushed here hasn't yet touched the existing database and thus 
should not effect business as usual. For now, in order to "run" the new 
process you have to go into `results/notebooks/datapackaging_prep`. Most 
of the new functions are in the new `ETL_pkg` module. Everything is 
still very experimental and not well documented.

Working on Issue #294