Reorganize repository for tox and python packaging
Lots of big changes here in support of our first release, including:
 * Moving the main `pudl` package into a `src` directory in order to isolate it
   from the packaging environment.
 * Moving all non-code files that are required for PUDL database ETL into a
   `data` directory within the python package for distribution. That directory
   is organized into several subdirectories, by data source for actual data
   that we need to distribute (`epa`, `eia`, `ferc1`, etc.) because it is not
   readily available, a `meta` directory for metadata, a `notebooks` folder
   for the tested example notebooks previously under `docs/notebooks` and a
   `settings` folder for the default YAML files that our scripts consume, and a
   `glue` folder for inter-datasource glue (like the eia923_ferc1 id_mapping
   spreadsheet)
 * Added the necessary details to the `setup.py` file to actually build an
   installable package.
 * Changed all references to files (e.g. the xlsx_maps in metadata) that were
   relative to the module `__file__` property to use the more general
   `pkg_resources` module which will provide reliable cross-platform access to
   the files distributed with the package, even when they are part of a binary
   distribution.
 * Set up a basic `tox` environment to test the python packaging and
   installation, as well as running pytest. Right now it is attempting to run
   on Python 3.6 and Python 3.7, since we expect PUDL to work on both of
   those versions. No major package versions from other dependencies are being
   used in the testing matrix.

Of these changes, the `tox` environment will probably have the biggest impact
on our overall usage.  Instead of running pytest directly, we will now use
`tox` both locally and on Travis (and other CI platforms if we set them up) to
run the various types of tests, in isolated packaged installation environments
so we know if the python packaging is working in addition to knowing if the
python code itself is working. The top level `tox.ini` file specifies the
various configured test environments. To run them you do:
```sh
tox -e testenv
```
where testenv can be one of the following:
 - `fast` (just run the quickie travis_ci tests, locally)
 - `travis` (run the travis tests for real, ONLY FOR USE ON TRAVIS CI)
 - `etl` (full etl, using the pudl_test database)
 - `validation` (post_etl data validation and notebooks)
 - `docs` (attempt to generate the full Sphinx documentation)

Note that none of this quite works yet, because the datastore and settings
files still need to be re-organized and disentangled from the `__file__`
attribute of the package -- basically the globally accessible PUDL_DIR needs
to be set somewhere so that the software knows where to put and look for data
and outputs, as outlined in #320.

Major progress toward #63 and #326
Closes: #331